+++ sound.c
@@ -1,6 +1,8 @@
 #include <alsa/asoundlib.h>
 #include "sound.h"
 #include "objects.h"
+#include "icon.h"
+#include "utils.h"
 
 static int success;
 static long max;
@@ -78,8 +80,35 @@ void sound_volume_down() {
 
 void sound_show(void) {
 	int volume = sound_get_volume();
-	char *text = malloc(sizeof(char) * 100);
-	sprintf(text, "Volume : %d%%", volume);
+	char *text = malloc(sizeof(char) * 5);
+	sprintf(text, "%d%%", volume);
 
-	XDrawString(display, window, gc_text_white, 10, 10, text, strlen(text));
+	char *sound_logo = utils_get(UTILS_SOUND);
+
+	Pixmap panel = XCreatePixmap(display, window, 200, 60, screen_depth);
+	Pixmap sound_pixmap = icon_get_pixmap(sound_logo, 40, 40);
+
+	free(sound_logo);
+
+	XFillRectangle(display, panel, gc_icon, 0, 0, 200, 60);
+	XCopyArea(display, sound_pixmap, panel, XDefaultGCOfScreen(screen), 0, 0, 40, 40, 5, 10);
+
+	const int width = 12 * (volume * 0.1);
+	XColor color_bar = {
+		.red = 0,
+		.green = 65535,
+		.blue = 0
+	};
+	XAllocColor(display, colormap, &color_bar);
+	XGCValues gcv_bar = {
+		.foreground = color_bar.pixel
+	};
+	GC gc_bar = XCreateGC(display, window, GCForeground, &gcv_bar);
+	XFillRectangle(display, panel, gc_bar, 50, 25, width, 10);
+
+	XDrawString(display, panel, gc_text_white, 170, 35, text, strlen(text));
+
+	const int x = (screen_width - 200) / 2;
+	const int y = (screen_height / 2);
+	XCopyArea(display, panel, window, XDefaultGCOfScreen(screen), 0, 0, 200, 60, x, y);
 }